fixed slicing: renamed types.py to custom_types.py
