+++ diffyscan/utils/custom_types.py
